match filename during skifile construction
When substituting the star / star forming region particle filenames, the stellar library name is used in the `lxml` xpath lookup. If one wants to use a different stellar library for a given component the search fails. 

This commit updates the search to match the `filename` field of the `SPHStellarComp` object, so that arbitrary stellar libraries can be used.